organizing functions for testing surface code
