+++ nilearn/surface/testing.py
@@ -1,0 +1,24 @@
+# Helper functions for testing surface code
+
+import numpy as np
+from scipy.spatial import Delaunay
+
+def _generate_surf():
+    rng = np.random.RandomState(42)
+    coords = rng.rand(20, 3)
+    faces = rng.randint(coords.shape[0], size=(30, 3))
+    return [coords, faces]
+
+def _flat_mesh(x_s, y_s, z=0):
+    x, y = np.mgrid[:x_s, :y_s]
+    x, y = x.ravel(), y.ravel()
+    z = np.ones(len(x)) * z
+    vertices = np.asarray([x, y, z]).T
+    triangulation = Delaunay(vertices[:, :2]).simplices
+    mesh = [vertices, triangulation]
+    return mesh
+
+
+def _z_const_img(x_s, y_s, z_s):
+    hslice = np.arange(x_s * y_s).reshape((x_s, y_s))
+    return np.ones((x_s, y_s, z_s)) * hslice[:, :, np.newaxis]
